% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.internal_fetch}
\alias{.internal_fetch}
\title{Internal fetch for non-exported functions}
\usage{
.internal_fetch(fname, pkg)
}
\arguments{
\item{fname}{Function name (string)}

\item{pkg}{Package name (string)}
}
\value{
The function object
}
\description{
Dynamically fetch a non-exported function from another namespace.
Used to access internal iNEXT.3D and iNEXT.beta3D functions safely.
}
