\name{esophagus}
\alias{esophagus}
\docType{data}
\title{
esophagus OTU count data
}
\description{
  This dataset is from package \code{phyloseq} and collected by Pei et al. (2004).
}
\usage{data(esophagus)}
\format{
  \code{esophagus} is a list with four species-by-assemblage data.frames. \cr
  $ esophagus_BC: \cr
  ..$ B: num [1:58] 50 0 2 0 0 ... \cr
  ..$ C: num [1:58] 19 2 13 2 0 ... \cr
  $ esophagus_BD: \cr
  ..$ B: num [1:58] 50 0 2 0 0 ... \cr
  ..$ D: num [1:58] 5 0 0 1 1 ... \cr
  $ esophagus_CD: \cr
  ..$ C: num [1:58] 19 2 13 2 0 ... \cr
  ..$ D: num [1:58] 5 0 0 1 1 ... \cr
  $ esophagus_BCD: \cr
  ..$ B: num [1:58] 50 0 2 0 0 ... \cr
  ..$ C: num [1:58] 19 2 13 2 0 ... \cr
  ..$ D: num [1:58] 5 0 0 1 1 ... \cr
 }
 
\references{
Pei, Z., Bini, E. J., Yang, L., Zhou, M., Francois, F., & Blaser, M. J. (2004). Bacterial biota in the human distal esophagus. Proceedings of the National Academy of Sciences of the United States of America, 101(12), 4250-4255. \cr\cr
}

\keyword{datasets}
