\name{esophagus_tree}
\alias{esophagus_tree}
\docType{data}
\title{
phylogenetic tree for \code{esophagus}
}
\description{
  This dataset includes a phylogenetic tree spanned by 58 OTU listed in the dataset \code{esophagus}. \cr
  Use the \code{chronos} function in the \code{ape} package to convert to an ultrametric tree with a fixed tree height of 1.
}
\usage{data(esophagus_tree)}
\format{
  \code{esophagus_tree} is a list with the following phylogenetic information: \cr
  A list of 4: \cr
  $ edge: int [1:114, 1:2] 59 59 60 60 61 62 63 63 62 61 ... \cr
  $ edge.length: num [1:114] 1 0.113 0.887 0.102 0.033 ... \cr
  $ Nnode: int 57 \cr
  $ tip.label: chr [1:58] "OTU_59_8_22" "OTU_59_5_13" "OTU_59_8_12" ... \cr
 }

\keyword{datasets}
