\name{fungi}
\alias{fungi}
\docType{data}
\title{Fungi OTU count data}

\description{
  This dataset is based on the study by V\'{e}trovsk\'{y} et al. (2015), published in the Dryad Digital Repository. \cr
  The dataset is titled \emph{``The rpb2 gene represents a viable alternative molecular marker for the analysis of environmental fungal communities.''}
}

\usage{data(fungi)}

\format{
A list of two species-by-assemblage \code{data.frame} objects: \cr

\itemize{
  \item \code{T03T04}: \cr
  \code{T03}, \code{T04}: Numeric vectors with OTU abundances (length 340) \cr

  \item \code{T09T10}: \cr
  \code{T09}, \code{T10}: Numeric vectors with OTU abundances (length 340) \cr
}
}

\references{
V\'{e}trovsk\'{y}, T., \\& Baldrian, P. (2015). Data from: The rpb2 gene represents a viable alternative molecular marker for the analysis of environmental fungal communities. Dryad Digital Repository. \doi{10.5061/dryad.vj3p9} \cr\cr
}
\keyword{datasets}
