% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ggiNEXT_seq_Relative}
\alias{ggiNEXT_seq_Relative}
\title{Plot Relative Gamma, Alpha, Beta or Dissimilarity from iNEXT_seq_Relative output}
\usage{
ggiNEXT_seq_Relative(output, type = "B")
}
\arguments{
\item{output}{Output list from \code{iNEXT_seq_Relative}}

\item{type}{Type of plot. "B" for diversity (gamma, alpha, beta), "D" for dissimilarity (1-C, 1-U, 1-V, 1-S).}
}
\value{
A ggplot object showing diversity or dissimilarity profiles.
}
\description{
This function creates ggplot-based plots for the output of \code{iNEXT_seq_Relative}.
}
\examples{
data("fungi")
data("fungi_tree")
output <- iNEXT_seq_Relative(fungi[1], q = c(0,1,2), nboot = 10, PDtree = fungi_tree)
ggiNEXT_seq_Relative(output, type = "B")
}
