% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ggObsAsyPD}
\alias{ggObsAsyPD}
\title{ggplot2 extension for an ObsAsyPD object}
\usage{
ggObsAsyPD(output, type = "B")
}
\arguments{
\item{output}{the output from ObsAsyPD.}

\item{method}{selection of plot type : \cr
\code{type = 'B'} for plotting the gamma, alpha, and beta diversity;  \cr
\code{type = 'D'} for plotting 4 turnover dissimilarities.}
}
\value{
a figure for phylogenetic diversity decomposition or dissimilarity measure.
}
\description{
\code{ggObsAsyPD}: the \code{\link[ggplot2]{ggplot}} extension for \code{\link{ObsAsyPD}} object to plot order q against to phylogenetic diversity decomposition and dissimilarity measure.
}
\examples{

data("esophagus")
data("esophagus_tree")
ObsAsyPD_output <- ObsAsyPD(esophagus[1], q = seq(0, 2, 0.2), weight = "size", nboot = 10, 
                           PDtree = esophagus_tree, type = "mle", decomposition = "relative")
ggObsAsyPD(ObsAsyPD_output, type = "B")

}
