\name{global_mat}
\alias{global_mat}
\docType{data}
\title{
hierarchical structure matrix for \code{global}
}
\description{
  This dataset includes a hierarchical structure matrix in the dataset \code{global} for calculate hierarchical diversity.
}
\usage{data(global_mat)}
\format{
  \code{global_mat} is a 3 x 7 (number of hierarchical layers times number of assemblages) matrix.
 }

\keyword{datasets}
