% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{hierPD}
\alias{hierPD}
\title{function to calculate hierarchical phylogenetic gamma, alpha, beta diversity and dissimilarity measures}
\usage{
hierPD(
  data,
  mat,
  q = seq(0, 2, 0.2),
  weight = "size",
  nboot = 10,
  conf = 0.95,
  PDtree,
  type = "mle",
  decomposition = "relative"
)
}
\arguments{
\item{data}{data should be input as a \code{matrix/data.frame} (species by assemblages).}

\item{mat}{hierarchical structure of data should be input as a \code{matrix}.}

\item{q}{a numerical vector specifying the diversity orders. Default is \code{seq(0, 2, 0.2)}.}

\item{weight}{(required only when \code{type = "mle"} and \code{decomposition = "relative"}) weight for relative decomposition empirical estimate. Select size-weighted \code{("size")}, equal-weighted \code{("equal")} or a numerical vector for weight. Default is \code{"size"}.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. Bootstrap replications are generally time consuming. Enter \code{0} to skip the bootstrap procedures. Default is \code{10}.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is \code{0.95}.}

\item{PDtree}{a phylogenetic tree in Newick format for all observed species in the pooled assemblage.}

\item{type}{estimate type: empirical \code{(type = "mle")} or asymptotic estimate \code{(type = "est")}. Default is \code{"mle"}.}

\item{decomposition}{decomposition type: relative \code{(decomposition = "relative")} or absolute decomposition \code{(decomposition = "absolute")}. Default is \code{"relative"}.}
}
\value{
a data frames with hierarchical phylogenetic diversity (gamma, alpha, and beta) and four types dissimilarity measure.
}
\description{
\code{hierPD}: function to calculate observed and asymptotic diversity estimates for hierarchical phylogenetic gamma, alpha, beta diversity and dissimilarity measures.
}
\examples{

data("global")
data("global_tree")
data("global_mat")
hier_output <- hierPD(global, mat = global_mat, q = seq(0, 2, 0.2), PDtree = global_tree)

hier_output
}
