% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ObsAsyPD}
\alias{ObsAsyPD}
\title{Asymptotic and observed phylogenetic gamma, alpha, beta diversity and dissimilarity of order q}
\usage{
ObsAsyPD(
  data,
  q = seq(0, 2, 0.2),
  weight = "size",
  nboot = 10,
  conf = 0.95,
  PDtree,
  type = c("mle", "est"),
  decomposition = c("absolute", "relative")
)
}
\arguments{
\item{data}{OTU data can be input as a \code{matrix/data.frame} (species by assemblages), or a \code{list} of \code{matrices/data.frames}, each matrix represents species-by-assemblages abundance matrix.\cr}

\item{q}{a numerical vector specifying the diversity orders. Default is \code{seq(0, 2, by = 0.2)}.}

\item{weight}{weight for relative decomposition. Default is \code{"size"}.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. Bootstrap replications are generally time consuming. Enter \code{0} to skip the bootstrap procedures. Default is \code{10}.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is 0.95.}

\item{PDtree}{a phylogenetic tree in Newick format for all observed species in the pooled assemblage.}

\item{type}{estimate type: estimate \code{(type = "est")}, empirical estimate \code{(type = "mle")}. Default is \code{"mle"}.}

\item{decomposition}{relative decomposition: \code{(decomposition = "relative")}, Absolute decomposition: \code{(decomposition = "absolute")}. Default is \code{"relative"}.}

\item{PDreftime}{a numerical value specifying reference time for PD. Default is \code{NULL} (i.e., the age of the root of PDtree).}
}
\value{
a data frames with asymptotic or observed phylogenetic diversity (gamma, alpha, and beta) and four types dissimilarity measure.
}
\description{
\code{ObsAsyPD} computes observed and asymptotic diversity of order q between 0 and 2 (in increments of 0.2) for phylogenetic gamma, alpha, beta diversity and dissimilarity; these values with different order q can be used to depict a q-profile in the \code{ggObsAsyPD} function.\cr\cr 
For each dimension, by default, both the observed and asymptotic diversity estimates will be computed.
}
\examples{

data("esophagus")
data("esophagus_tree")
ObsAsyPD_output <- ObsAsyPD(esophagus$BCD, q = seq(0, 2, 0.2), weight = "size", nboot = 10, 
                           PDtree = esophagus_tree, type = "mle", decomposition = "relative")

}
