\name{tongue_cheek}
\alias{tongue_cheek}
\docType{data}
\title{
Tongue and Cheek OTU Count Data
}
\description{
A dataset containing OTU (Operational Taxonomic Unit) abundances from human cheek and tongue samples.
}
\usage{data("tongue_cheek")}
\format{
  A data frame with 2549 observations on the following 2 variables.
  \describe{
    \item{\code{Cheek}}{a numeric vector representing OTU abundances from cheek samples.}
    \item{\code{Tongue}}{a numeric vector representing OTU abundances from tongue samples.}
  }
}
\details{
This dataset can be used to explore phylogenetic diversity and dissimilarity between oral microbial communities sampled from the cheek and the tongue.
}
\source{
Simulated dataset for demonstration purposes.
}
\examples{
data(tongue_cheek)
}
\keyword{datasets}
