\name{wetland}
\alias{wetland}
\docType{data}
\title{
wetland OTU count data
}
\description{
  This dataset is from package \code{microeco} and collected by An et al. (2019).
}
\usage{data(wetland)}
\format{
  \code{wetland} is a species-by-assemblages data.frame. \cr
  $ NE: num 132 12 15 14 12 ... \cr
  $ NW: num 10 4 0 4 7 9 ... \cr
  $ NC: num 17 9 1 24 26 ... \cr
  $ YML: num 0 1 0 10 2 ... \cr
  $ SC: num 32 37 18 23 153 ... \cr
 }
 
\references{
An, J., Liu, C., Wang, Q., Yao, M., Rui, J., Zhang, S., & Li, X. (2019). Soil bacterial community structure in Chinese wetlands. Geoderma, 337, 290-299. \cr\cr
}

\keyword{datasets}
