\name{wetland_mat}
\alias{wetland_mat}
\docType{data}
\title{
hierarchical structure matrix for \code{wetland}
}
\description{
  This dataset includes a hierarchical structure matrix in the dataset \code{wetland} for calculate hierarchical diversity.
}
\usage{data(wetland_mat)}
\format{
  \code{wetland_mat} is a 3 x 5 (number of hierarchical layers times number of assemblages) matrix.
 }

\keyword{datasets}
