\name{wetland_tree}
\alias{wetland_tree}
\docType{data}
\title{
phylogenetic tree for \code{wetland}
}
\description{
  This dataset includes a phylogenetic tree spanned by 404 OTU listed in the dataset \code{wetland}. \cr
  Use the \code{chronos} function in the \code{ape} package to convert to an ultrametric tree with a fixed tree height of 1.
}
\usage{data(wetland_tree)}
\format{
  \code{wetland_tree} is a list with the following phylogenetic information: \cr
  A list of 4: \cr
  $ edge: int [1:806, 1:2] 405 406 407 408 409 410 410 411 412 412 ... \cr
  $ edge.length: num [1:806] 9.99e-16 3.86e-01 5.19e-02 2.13e-01 2.89e-02 ... \cr
  $ Nnode: int 403 \cr
  $ tip.label: chr [1:404] "OTU_3829" "OTU_705" "OTU_3570" ... \cr
 }

\keyword{datasets}
