% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{gghierPD}
\alias{gghierPD}
\title{ggplot2 extension for an hierPD object}
\usage{
gghierPD(output, method = "A")
}
\arguments{
\item{output}{the output from hierPD.}

\item{method}{\code{(method = "A")} diversity(alpha, gamma); \code{(method = "B")} beta diversity; \code{(method = "D")} dissimilarity measure based on multiplicative decomposition.}
}
\value{
a figure for hierarchical phylogenetic diversity decomposition or dissimilarity measure.
}
\description{
\code{gghierPD}: the \code{\link[ggplot2]{ggplot}} extension for \code{\link{hierPD}} object to plot order q against to hierarchical phylogenetic diversity decomposition and dissimilarity measure.
}
\examples{

data("antechinus")
data("antechinus_mat")
data("antechinus_tree")
hier_output <- hierPD(antechinus, mat = antechinus_mat, tree = antechinus_tree, q = seq(0, 2, 0.2))
gghierPD(hier_output, method = "A")

}
