% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{gghierPD}
\alias{gghierPD}
\title{ggplot2 extension for an hierPD object}
\usage{
gghierPD(output, type = "A")
}
\arguments{
\item{output}{the output from hierPD.}

\item{type}{selection of plot type : \cr
\code{(type = "A")} for alpha and gamma diversity; \cr
\code{(type = "B")} for beta diversity; \cr
\code{(type = "D")} for dissimilarity measure based on multiplicative decomposition.}
}
\value{
a figure for hierarchical phylogenetic diversity decomposition or dissimilarity measure.
}
\description{
\code{gghierPD}: the \code{ggplot} extension for \code{\link{hierPD}} object to plot order q against to hierarchical phylogenetic diversity decomposition and dissimilarity measure.
}
\examples{

data("wetland")
data("wetland_mat")
data("wetland_tree")
hier_output <- hierPD(wetland, mat = wetland_mat, q = seq(0, 2, 0.2), PDtree = wetland_tree)
gghierPD(hier_output, type = "A")

}
