% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ggiNEXTseq}
\alias{ggiNEXTseq}
\title{ggplot2 extension for an iNEXT.seq object}
\usage{
ggiNEXTseq(output, type = "B")
}
\arguments{
\item{output}{the output from iNEXTseq.}

\item{type}{(required only when \code{base = "coverage"}), selection of plot type : \cr
\code{type = 'B'} for plotting the gamma, alpha, and beta diversity;  \cr
\code{type = 'D'} for plotting 4 turnover dissimilarities.}
}
\value{
a figure for phylogenetic diversity decomposition or dissimilarity measure.
}
\description{
\code{ggiNEXTseq}: the \code{\link[ggplot2]{ggplot}} extension for \code{\link{iNEXTseq}} object to plot coverage- or sample-sized-based rarefaction/extrapolation curves for phylogenetic diversity decomposition and dissimilarity measure.
}
\examples{

data("esophagus")
data("esophagus_tree")
output <- iNEXTseq(esophagus[1], q = c(0,1,2), nboot = 10, PDtree = esophagus_tree)
ggiNEXTseq(output, type = "B")

}
