% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ggObsAsyPD}
\alias{ggObsAsyPD}
\title{ggplot2 extension for an ObsAsyPD object}
\usage{
ggObsAsyPD(output, type = "B")
}
\arguments{
\item{output}{The output from \code{\link{ObsAsyPD}}.}

\item{type}{Selection of plot type: \cr
\code{"B"} for plotting gamma, alpha, and beta diversity; \cr
\code{"D"} for plotting four turnover dissimilarities.}
}
\value{
A ggplot object showing phylogenetic diversity or dissimilarity.
}
\description{
\code{ggObsAsyPD}: the \code{ggplot} extension for \code{\link{ObsAsyPD}} object to plot order q against phylogenetic diversity decomposition or dissimilarity measure.
}
\examples{
data("fungi")
data("fungi_tree")
ObsAsyPD_output <- ObsAsyPD(fungi[1], q = seq(0, 2, 0.2), PDtree = fungi_tree)
ggObsAsyPD(ObsAsyPD_output, type = "B")

}
