% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{hierPD}
\alias{hierPD}
\title{function to calculate hierarchical phylogenetic gamma, alpha, beta diversity and dissimilarity measure}
\usage{
hierPD(
  data,
  mat,
  PDtree,
  q = seq(0, 2, 0.2),
  weight = "size",
  nboot = 20,
  conf = 0.95,
  type = "mle",
  decomposition = "relative"
)
}
\arguments{
\item{data}{data should be input as a \code{matrix/data.frame} (species by assemblages).}

\item{mat}{hierarchical structure of data should be input as a \code{matrix}.}

\item{q}{a numerical vector specifying the diversity orders. Default is \code{seq(0, 2, 0.2)}.}

\item{weight}{weight for relative decomposition. Default is \code{"size"}.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. Bootstrap replications are generally time consuming. Enter \code{0} to skip the bootstrap procedures. Default is \code{20}.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is \code{0.95}.}

\item{type}{estimate type: estimate \code{(type = "est")}, empirical estimate \code{(type = "mle")}. Default is \code{"mle"}.}

\item{decomposition}{relative decomposition: \code{(decomposition = "relative")}, Absolute decomposition: \code{(decomposition = "absolute")}.}

\item{tree}{a phylogenetic tree in Newick format for all observed species in the pooled assemblage.}
}
\value{
a data frames with hierarchical phylogenetic diversity (gamma, alpha, and beta) and four types dissimilarity measure.
}
\description{
\code{hierPD}: function to calculate empirical estimates for hierarchical phylogenetic gamma, alpha, beta diversity and dissimilarity measure
}
\examples{

data("antechinus")
data("antechinus_mat")
data("antechinus_tree")
hier_output <- hierPD(antechinus, mat = antechinus_mat, tree = antechinus_tree, q = seq(0, 2, 0.2))

}
