% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{iNEXTseq}
\alias{iNEXTseq}
\title{function to calculate phylogenetic gamma, alpha, beta diversity and dissimilarity measure}
\usage{
iNEXTseq(
  data,
  q = c(0, 1, 2),
  base = "coverage",
  level = NULL,
  nboot = 10,
  conf = 0.95,
  PDtree = NULL,
  PDreftime = NULL
)
}
\arguments{
\item{data}{OTU data can be input as a \code{matrix/data.frame} (species by assemblages), or a \code{list} of \code{matrices/data.frames}, each matrix represents species-by-assemblages abundance matrix.\cr}

\item{q}{a numerical vector specifying the diversity orders. Default is c(0, 1, 2).}

\item{base}{sample-sized-based rarefaction and extrapolation for gamma and alpha diversity (\code{base = "size"}) or coverage-based rarefaction and extrapolation for gamma, alpha and beta diversity (\code{base = "coverage"}). Default is \code{base = "coverage"}.}

\item{level}{A numerical vector specifying the particular value of sample coverage (between 0 and 1 when \code{base = "coverage"}) or sample size (\code{base = "size"}). \code{level = 1} (\code{base = "coverage"}) means complete coverage (the corresponding diversity represents asymptotic diversity).\cr
If \code{base = "size"} and \code{level = NULL}, then this function computes the gamma and alpha diversity estimates up to double the reference sample size. \cr
If \code{base = "coverage"} and \code{level = NULL}, then this function computes the gamma and alpha diversity estimates up to one (for \code{q = 1, 2}) or up to the coverage of double the reference sample size (for \code{q = 0});
the corresponding beta diversity is computed up to the same maximum coverage as the alpha diversity.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. Bootstrap replications are generally time consuming. Enter \code{0} to skip the bootstrap procedures. Default is \code{10}.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is \code{0.95}.}

\item{PDtree}{a phylogenetic tree in Newick format for all observed species in the pooled assemblage.}

\item{PDreftime}{a numerical value specifying reference time for PD. Default is \code{NULL} (i.e., the age of the root of PDtree).}
}
\value{
If \code{base = "coverage"}, return a list of seven data frames with three coverage-based diversity (gamma, alpha, and beta) and four types dissimilarity measure. If \code{base = "size"}, return a list of two data frames with two diversity (gamma and alpha).
}
\description{
\code{iNEXTseq}: function to calculate interpolation and extrapolation for phylogenetic gamma, alpha, beta diversity and dissimilarity measure
}
\examples{

data("tongue_cheek")
data("tongue_cheek_tree")
output <- iNEXTseq(tongue_cheek, q=c(0,1,2),
                       level = seq(0.5, 1, 0.05), nboot = 10,
                       conf = 0.95, PDtree = tongue_cheek_tree, PDreftime = NULL)

}
