% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ObsAsyPD}
\alias{ObsAsyPD}
\title{function to calculate observed or asymptotic phylogenetic gamma, alpha, beta diversity and dissimilarity of order q}
\usage{
ObsAsyPD(
  data,
  q = seq(0, 2, 0.2),
  weight = "size",
  nboot = 10,
  conf = 0.95,
  PDtree,
  type = "mle",
  decomposition = "relative"
)
}
\arguments{
\item{data}{OTU data can be input as a \code{matrix/data.frame} (species by assemblages), or a \code{list} of \code{matrices/data.frames}, each matrix represents species-by-assemblages abundance matrix.\cr}

\item{q}{a numerical vector specifying the diversity orders. Default is \code{seq(0, 2, 0.2)}.}

\item{weight}{(required only when \code{type = "mle"} and \code{decomposition = "relative"}) weight for relative decomposition empirical estimate. Select size-weighted \code{("size")}, equal-weighted \code{("equal")} or a numerical vector for weight. Default is \code{"size"}.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. Bootstrap replications are generally time consuming. Enter \code{0} to skip the bootstrap procedures. Default is \code{10}.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is \code{0.95}.}

\item{PDtree}{a phylogenetic tree in Newick format for all observed species in the pooled assemblage.}

\item{type}{estimate type: empirical \code{(type = "mle")} or asymptotic estimate \code{(type = "est")}. Default is \code{"mle"}.}

\item{decomposition}{decomposition type: relative \code{(decomposition = "relative")} or absolute decomposition \code{(decomposition = "absolute")}. Default is \code{"relative"}.}
}
\value{
a data frames with observed or asymptotic phylogenetic diversity (gamma, alpha, and beta) and four types dissimilarity measure for each dataset.
}
\description{
\code{ObsAsyPD} computes observed and asymptotic diversity of order q between 0 and 2 (in increments of 0.2) for phylogenetic gamma, alpha, beta diversity and dissimilarity; these values with different order q can be used to depict a q-profile in the \code{ggObsAsyPD} function.
}
\examples{

data("fungi")
data("fungi_tree")
ObsAsyPD_output <- ObsAsyPD(fungi[1], q = seq(0, 2, 0.2), PDtree = fungi_tree)

ObsAsyPD_output
}
